reorganized register req and replaced sanitization with schema
